Don't truncate the action's pseudo path names before we give them to GetLock()
GetLock() does a sort of checksum over the path names if they're too
long and makes sure that they fit inside the allocated buffer

So, if we pre-truncate long path names there's nothing their to allow
them to uniquify with in GetLock().

Later on, if another operation (eg: copy) takes long enough, the
GetLastLock() elapsedtime check will go negative and it will appear as
though multiple processes have already performed a link operation if
their long names collide on the (last) lock name.

That's in fact due to another bug in terms of which timestamp we're
comparing the last lock to - CFSTARTTIME (the time cfagent started) or
CFINITSTARTTIME (the time the last pass started).

Really, it should be compared against the first time that a given action
was first run, but we'll leave that for another commit.